refactor(backend): TSK-02 Implement Flask application factory pattern and add wsgi.py for migrations
