--- app/run.py
@@ -1,5 +1,0 @@
-from app import app,
-
-if __name__ == '__main__':
-    app.run(port=5555, debug=True)
-
